[ALPS03186560] TEEI: RPMB
[Detail]
Modify RPMB and add TUI.

MTK-Commit-Id: 56ae080c2deef8b33fa5872176b96a2f00efa681

CR-Id: ALPS03186560
Feature: Microtrust TEEI
(cherry picked from commit 062c19f62be492920f6e855853d855585e5f438a)

+++ drivers/misc/mediatek/teei/V1.0/tz_driver/include/teei_common.h
@@ -69,7 +69,7 @@ enum teeic_param_pos {
 struct teei_encode_meta {
 	int type;
 	int len;                        /* data length */
-	unsigned  long long usr_addr;           /* data address in user space */
+	unsigned long usr_addr;		/* data address in user space */
 	int ret_len;                    /* return sizeof data */
 	int value_flag;                 /* value of a or b */
 	int param_pos;                  /* param order */
